requirements.txt
# Needed packages/versions
matplotlib
gspread
ipdb
numpy
pandas
requests
google.cloud
oauth2client